for job reference, use the runner of child steps
